Moved adding deprecatedId to EditedFormFinalizationProcessor
